requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
django-filter>=24.2
mysqlclient>=2.2
Pillow>=10.0
markdown2>=2.4
